fix: prevent login error when using module slot
If a user has only Public (module protected) objects they can
choose the module slot as their token. The new v1.0 API now tries
to login regardless if the user provided a pin. This will return
an error when using the module slot as it does not require a login.

I don't know the full intent of the original change to the API but
I figured it would be better to continue to call Login and just ignore
the error to not break current behavior.